requirements.txt
psycopg2-binary
pytest
fastapi